Refactor coin model and methods out into coin class module
